requirements.txt
# Core dependencies
pyspark>=3.5.0
findspark>=2.0.1
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=10.0.0

# Visualization (for notebooks)
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Development dependencies
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
